+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Unreleased
 
 - Added "sample" linking to `log_viewer()` data preparation function.
-- Allow custom `http_client` to OpenAI clients.
+- Allow custom `http_client` to OpenAI clients
+- Added sample, message, and event linking to `log_viewer()` data preparation function.
 
 ## 0.3.116 (27 July 2025)